requirements.txt
ibm-watson
python-dotenv
pandas
numpy
